ai/server.py
"""
Orbital Eye — AI Worker (gRPC Server)

Handles all ML inference: object detection, classification, change detection.
Called from the Go main process via gRPC.
"""

import argparse
import logging
from concurrent import futures

import grpc

# Generated proto stubs (run: python -m grpc_tools.protoc ...)
# from proto.gen import detector_pb2, detector_pb2_grpc

logging.basicConfig(level=logging.INFO, format="%(asctime)s [%(levelname)s] %(message)s")
logger = logging.getLogger("ai-worker")


class DetectorServicer:
    """Implements the DetectorService gRPC interface."""

    def __init__(self, model_dir: str = "models/", device: str = "auto"):
        self.model_dir = model_dir
        self.device = device
        self.models = {}
        logger.info(f"AI Worker initialized (device={device}, models={model_dir})")

    def _load_model(self, name: str):
        """Lazy-load a detection model."""
        if name in self.models:
            return self.models[name]

        logger.info(f"Loading model: {name}")
        # TODO: Load from HuggingFace Hub or local path
        # For vessel detection: ultralytics YOLO with custom weights
        # For change detection: Siamese UNet
        # For classification: fine-tuned ResNet/EfficientNet
        raise NotImplementedError(f"Model {name} not yet available")

    def DetectObjects(self, request, context):
        """Run object detection on a satellite image tile."""
        logger.info(f"DetectObjects: targets={request.target_classes}, gsd={request.gsd_meters}m")
        # TODO: implement
        # 1. Load/decode image
        # 2. Preprocess (normalize, tile if needed)
        # 3. Run YOLO inference
        # 4. Convert pixel coords → geo coords using GSD + reference point
        # 5. Return detections
        return None  # DetectResponse

    def ClassifyObject(self, request, context):
        """Fine-grained classification of a detected object."""
        logger.info(f"ClassifyObject: coarse_class={request.coarse_class}")
        # TODO: implement
        return None

    def DetectChanges(self, request, context):
        """Compare two temporal images for changes."""
        logger.info(f"DetectChanges: sensitivity={request.sensitivity}")
        # TODO: implement
        # 1. Co-register images
        # 2. Run change detection model
        # 3. Filter by significance
        return None

    def Enhance(self, request, context):
        """Super-resolution enhancement."""
        logger.info(f"Enhance: scale={request.scale_factor}x")
        # TODO: implement
        return None

    def Health(self, request, context):
        """Health check."""
        import torch
        gpu_used, gpu_total = 0, 0
        if torch.cuda.is_available():
            gpu_used = torch.cuda.memory_allocated() // (1024 * 1024)
            gpu_total = torch.cuda.get_device_properties(0).total_mem // (1024 * 1024)
        # TODO: return HealthResponse
        return None


def serve(port: int = 50051, model_dir: str = "models/", device: str = "auto"):
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=4))
    servicer = DetectorServicer(model_dir=model_dir, device=device)
    # detector_pb2_grpc.add_DetectorServiceServicer_to_server(servicer, server)
    server.add_insecure_port(f"[::]:{port}")
    server.start()
    logger.info(f"AI Worker listening on port {port}")
    server.wait_for_termination()


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Orbital Eye AI Worker")
    parser.add_argument("--port", type=int, default=50051)
    parser.add_argument("--model-dir", default="models/")
    parser.add_argument("--device", default="auto", help="cuda, cpu, or auto")
    args = parser.parse_args()
    serve(port=args.port, model_dir=args.model_dir, device=args.device)
