ai/requirements.txt
grpcio>=1.70.0
grpcio-tools>=1.70.0
protobuf>=5.29.0
torch>=2.0
torchvision>=0.15
ultralytics>=8.0
opencv-python-headless>=4.8
numpy>=1.24
pillow>=10.0
rasterio>=1.3
albumentations>=1.4
huggingface-hub>=0.27
